# openapi.yaml
openapi: 3.1.0
info:
  title: EasyMod API
  version: 1.0.0
  description: EasyMod Backend API
servers:
  - url: /api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /customer:
    parameters:
      - $ref: '#/components/parameters/XShopId'
      - $ref: '#/components/parameters/RateLimitInfo'
    get:
      tags: [Customer]
      summary: List customers
      responses:
        '200': { description: Customers }
    post:
      tags: [Customer]
      summary: Create customer
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: Alice
              email: alice@example.com
              phone: "+1-555-0101"
            schema:
              type: object
      responses:
        '201': { description: Created }
  /customer/{id}:
    parameters:
      - $ref: '#/components/parameters/XShopId'
      - $ref: '#/components/parameters/RateLimitInfo'
    get:
      tags: [Customer]
      summary: Get customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Customer }
    patch:
      tags: [Customer]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            example:
              phone: "+1-555-0102"
            schema:
              type: object
      responses:
        '200': { description: Updated }
    delete:
      tags: [Customer]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Deleted }
  /customer/list:
    parameters:
      - $ref: '#/components/parameters/XShopId'
      - $ref: '#/components/parameters/RateLimitInfo'
    get:
      tags: [Customer]
      summary: Legacy list customers
      responses:
        '200': { description: Customers }
  /customer/get:
    parameters:
      - $ref: '#/components/parameters/XShopId'
      - $ref: '#/components/parameters/RateLimitInfo'
    get:
      tags: [Customer]
      summary: Legacy get customer
      responses:
        '200': { description: Customer }
  /customer/create:
    parameters:
      - $ref: '#/components/parameters/XShopId'
      - $ref: '#/components/parameters/RateLimitInfo'
    post:
      tags: [Customer]
      summary: Legacy create customer
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            example:
              name: Alice
              email: alice@example.com
            schema:
              type: object
      responses:
        '201': { description: Created }
  /customer/update:
    parameters:
      - $ref: '#/components/parameters/XShopId'
      - $ref: '#/components/parameters/RateLimitInfo'
    post:
      tags: [Customer]
      summary: Legacy update customer
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            example:
              id: cust_123
              phone: "+1-555-0102"
            schema:
              type: object
      responses:
        '200': { description: Updated }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XShopId:
      name: x-shop-id
      in: header
      required: false
      schema:
        type: string
      description: Shop identifier for multi-tenant context
    RateLimitInfo:
      name: x-rate-limit-info
      in: header
      required: false
      schema:
        type: string
      description: Rate limit information header
    IdempotencyKey:
      name: idempotency-key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key for safe retries
...